[rotorcraft] move guided mode in a separated place
no reason to keep this mixed with the main autopilot logic and it should
help to have this more modular in the end
