+++ sw/airborne/firmwares/rotorcraft/rotorcraft_datalink.c
@@ -35,6 +35,7 @@
 #endif
 
 #include "firmwares/rotorcraft/autopilot.h"
+#include "firmwares/rotorcraft/autopilot_guided.h"
 
 void firmware_parse_msg(struct link_device *dev __attribute__((unused)), struct transport_tx *trans __attribute__((unused)), uint8_t *buf)
 {
